pyproject.toml
[tool.poetry]
name = "cherryservers"
version = "1.0.0"
description = "Cherry Servers Ansible Collection"
license = "GPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
ansible-core = "^2.14"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"